refactor: extract queue commands
